requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Training and optimization
torchtune>=0.1.0
torchao>=0.1.0

# Data processing
langdetect>=1.0.9

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For HDF5 dataset storage
h5py>=3.9.0

# Optional: For async monitoring
watchdog>=3.0.0

